requirements.txt
flask
flask-cors
xhtml2pdf
python-dotenv
openai